requirements.txt
Streamlit
numpy
Pillow
requests
opencv-python
FastAPI
uvicorn
pydantic
python-multipart
torch 
torchvision 
torchaudio
diffusers
ipywidgets
accelerate
transformers
safetensors
matplotlib
pycocotools
dotenv